Improved fixtures
